Improves button color and hover states
Refactors button styling to use slotted elements and CSS classes for color variations.
This approach provides better encapsulation and allows for dynamic color updates.

Adds hover states using JavaScript to manage the 'is-hovered' class,
and manages the `is-disabled` class based on the disabled state of the slotted element.

Renames the stylesheet to `template.scss` for clarity.